+++ src/Bet.Extensions.Http.MessageHandlers/Authorize/AuthType.cs
@@ -1,0 +1,8 @@
+﻿namespace Bet.Extensions.Http.MessageHandlers.Authorize
+{
+    public enum AuthType
+    {
+        Basic,
+        Bearer
+    }
+}