+++ src/Bet.Extensions.Resilience.Abstractions/PolicyShapes{TResult}.cs
@@ -140,7 +140,7 @@ namespace Polly
                 // failed bulkhead policy
                 .Or<BulkheadRejectedException>()
 
-                .Fallback<TResult>(fallbackAction: Fallback, onFallback: OnFallbackAction)
+                .Fallback(fallbackAction: Fallback, onFallback: OnFallbackAction)
 
                 .WithPolicyKey(options.Name);