+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 Change Log
 ===============================================================================
 
+Version 2.1.1
+----------------------------
+ * Update to DotNetCore 3.1.4
+ * Added support preview feed  `https://f.feedz.io/kdcllc/bet-extensions-resilience/nuget/index.json`
+ * Package was deprecated `Bet.Extensions.Resilience.SqlClient` in favor of `Bet.Extensions.Resilience.Data.SqlClient`
+ * `Bet.Extensions.Resilience.Data.Abstractions` package was introduced for `DbConnection`
+
 Version 1.0.0
 ----------------------------
  * Intial design